+++ Káoszrubin/Domain/Combat/WeaponDefinition.cs
@@ -4,7 +4,7 @@ namespace KaoszRubin.Domain.Combat;
 
 public sealed record WeaponDefinition(string Id, string Name, string? WeaponTypeId, ValueRange? Damage,
     int MinimumStrength, bool IsTwoHanded, IReadOnlySet<string> AllowedClassIds, string Description, int BasePrice,
-    ItemRarity Rarity = ItemRarity.Normal, string? BaseWeaponId = null, int MagicPower = 0, int Weight = 1) : IItemDefinition
+    ItemRarity Rarity = ItemRarity.Normal, string? BaseWeaponId = null, int MagicPower = 0, double Weight = 1) : IItemDefinition
 {
     public ItemCategory Category => ItemCategory.Weapon;
     public bool CanBeEquippedBy(string characterClassId, int strength) =>
